requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
blas=1.0=openblas
brotli=1.0.9=h1a28f6b_7
brotli-bin=1.0.9=h1a28f6b_7
brotlipy=0.7.0=py39h1a28f6b_1002
bzip2=1.0.8=h620ffc9_4
ca-certificates=2022.10.11=hca03da5_0
certifi=2022.12.7=py39hca03da5_0
cffi=1.15.1=py39h80987f9_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
contourpy=1.0.5=py39h525c30c_0
cryptography=38.0.4=py39h834c97f_0
cycler=0.11.0=pyhd3eb1b0_0
facenet-pytorch=2.5.2=pypi_0
ffmpeg=4.2.2=h04105a8_0
fftw=3.3.9=h1a28f6b_1
flit-core=3.6.0=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h1192e45_0
gettext=0.21.0=h826f4ad_0
giflib=5.2.1=h1a28f6b_0
gmp=6.2.1=hc377ac9_3
gnutls=3.6.15=h887c41c_0
icu=68.1=hc377ac9_0
idna=3.4=py39hca03da5_0
jpeg=9e=h1a28f6b_0
keras=2.10.0=py39hca03da5_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.4.4=py39h313beb8_0
lame=3.100=h1a28f6b_0
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libbrotlicommon=1.0.9=h1a28f6b_7
libbrotlidec=1.0.9=h1a28f6b_7
libbrotlienc=1.0.9=h1a28f6b_7
libcxx=14.0.6=h848a8c0_0
libdeflate=1.8=h1a28f6b_5
libffi=3.4.2=hca03da5_6
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libiconv=1.16=h1a28f6b_2
libidn2=2.3.1=h1a28f6b_0
libopenblas=0.3.21=h269037a_0
libopus=1.3=h1a28f6b_1
libpng=1.6.37=hb8d0fd4_0
libtasn1=4.16.0=h1a28f6b_0
libtiff=4.5.0=h2fd578a_0
libunistring=0.9.10=h1a28f6b_0
libvpx=1.10.0=hc377ac9_0
libwebp=1.2.4=h68602c7_0
libwebp-base=1.2.4=h1a28f6b_0
libxml2=2.9.14=h8c5e841_0
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_0
matplotlib=3.6.2=py39hca03da5_0
matplotlib-base=3.6.2=py39h8bbb115_0
munkres=1.1.4=py_0
ncurses=6.3=h1a28f6b_3
nettle=3.7.3=h84b5d62_1
numpy=1.23.5=py39h1398885_0
numpy-base=1.23.5=py39h90707a3_0
openh264=1.8.0=h98b2900_0
openssl=1.1.1s=h1a28f6b_0
packaging=22.0=py39hca03da5_0
pillow=9.3.0=py39hf4a492f_1
pip=22.3.1=py39hca03da5_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py39hca03da5_0
pysocks=1.7.1=py39hca03da5_0
python=3.9.15=hc0d8a6c_2
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.13.1=py3.9_0
readline=8.2=h1a28f6b_0
requests=2.28.1=py39hca03da5_0
scipy=1.9.3=py39h9d039d2_0
setuptools=65.6.3=py39hca03da5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.40.1=h7a7dc30_0
tk=8.6.12=hb8d0fd4_0
torchvision=0.14.1=py39_cpu
tornado=6.2=py39h1a28f6b_0
typing_extensions=4.4.0=py39hca03da5_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.13=py39hca03da5_0
wheel=0.37.1=pyhd3eb1b0_0
x264=1!152.20180806=h1a28f6b_0
xz=5.2.8=h80987f9_0
zlib=1.2.13=h5a0b063_0
zstd=1.5.2=h8574219_0
